Crate enum_set2

A set for enum variants

It is implemented as a wrapper over the bit-set crate, which provides a set for integers values. We use the EnumLike trait from the enum_like crate which allows for a conversion between enum variant and integer value.

Since an EnumSet is a wrapper over a BitSet, and a BitSet is a wrapper over a BitVec, looking at the assembly generated by this crate should be interesting.

For usage examples, check out https://github.com/Badel2/enum_vec/blob/master/example/src/enum_set.rs

Structs

EnumSet
A BitSet indexed by an EnumLike type.

WrapIter
Wraps an iterator from the bit-set crate, mapping the output from usize to E: EnumLike.

Type Definitions

Difference
Iterator over the &EnumSet

Intersection
Iterator over the &EnumSet

Iter
Iterator over the &EnumSet

SymmetricDifference
Iterator over the &EnumSet

Union
Iterator over the &EnumSet