Struct enum_set2::EnumSet

pub struct EnumSet<E: EnumLike> { /* fields omitted */ }

A BitSet indexed by an EnumLike type.

Methods

impl<E: EnumLike> EnumSet<E>

pub fn into_bit_set(self) -> BitSet

Returns the inner BitSet.

pub fn get_ref(&self) -> &BitSet

Returns a reference to the inner BitSet.

pub fn from_bit_set(inner: BitSet) -> Self

Constructs an EnumSet from a BitSet.

pub fn new() -> Self

Creates a new EnumSet.

pub fn shrink_to_fit(&mut self)

Attemps to minimalize the memory usage of the inner BitSet.

Important traits for WrapIter<E, I>
impl<E: EnumLike, I: Iterator<Item = usize>> Iterator for WrapIter<E, I> type Item = E;

pub fn iter(&self) -> WrapIter<E, Iter<u32>>

Iterator over each element in the set.

pub fn union<'a>(&'a self, other: &'a Self) -> Union<'a, E>

Iterator over each element in set || other.

pub fn intersection<'a>(&'a self, other: &'a Self) -> Intersection<'a, E>

Iterator over each element in set && other.

pub fn difference<'a>(&'a self, other: &'a Self) -> Difference<'a, E>

Iterator over each element in set - other.

pub fn symmetric_difference<'a>(
&'a self,
other: &'a Self
) -> SymmetricDifference<'a, E>

Iterator over each element in set ^ other.

pub fn union_with(&mut self, other: &Self)

Computes the union with the other set, in-place.

pub fn intersect_with(&mut self, other: &Self)

Computes the intersection with the other set, in-place.

pub fn difference_with(&mut self, other: &Self)

Computes the difference with the other set, in-place.

pub fn symmetric_difference_with(&mut self, other: &Self)

Computes the symmetric difference with the other set, in-place.

pub fn len(&self) -> usize

Returns the number of elements in the set.

pub fn is_empty(&self) -> bool

Returns true if there are no elements in the set.

pub fn clear(&mut self)

Clears all elements in the set.

pub fn contains(&self, value: E) -> bool

Returns true if the set contains this value.

pub fn is_disjoint(&self, other: &Self) -> bool

Returns true if the set has no elements in common with other.

pub fn is_subset(&self, other: &Self) -> bool

Returns true if the set has no elements in common with other.

pub fn is_superset(&self, other: &Self) -> bool

Returns true if the set has no elements in common with other.

pub fn insert(&mut self, value: E) -> bool

Returns true if the value was not already present in the set

pub fn remove(&mut self, value: E) -> bool

Returns true if the value was already present in the set